refactor: move leaderboard contents
