+++ packages/web/src/layouts/leaderboard-layout/Leaderboard.tsx
@@ -2,9 +2,9 @@ import React from "react";
 
 import Footer from "@components/common/footer/Footer";
 import HeaderContainer from "@containers/header-container/HeaderContainer";
-import LeaderboardSubHeaderContainer from "@containers/leaderboard-subheader-container/LeaderboardSubheaderContainer";
 
-import LeaderboardList from "./components/leaderboard-list-layout/LeaderboardList";
+import LeaderboardSubHeaderContainer from "./containers/leaderboard-subheader-container/LeaderboardSubheaderContainer";
+import LeaderboardList from "./leaderboard-list/LeaderboardList";
 import LeaderboardLayout from "./LeaderboardLayout";
 
 const Leaderboard: React.FC = () => {
