+++ packages/web/src/layouts/leaderboard-layout/components/common/common.styles.ts
