+++ packages/web/src/layouts/leaderboard-layout/components/common/styled-icon-file/StyledIconFile.tsx
