+++ packages/web/src/layouts/leaderboard-layout/components/common/styled-icon-info/StyledIconInfo.tsx
