+++ packages/web/src/layouts/leaderboard-layout/components/copy-referral-link/CopyReferralLink.stories.tsx
