+++ packages/web/src/layouts/leaderboard-layout/components/leaderboard-subheader/LeaderboardSubheader.styles.ts
