+++ packages/web/src/layouts/leaderboard-layout/components/leaderboard-table-header/LeaderboardTableHeader.styles.ts
