+++ packages/web/src/layouts/leaderboard-layout/components/leaderboard-table-header/LeaderboardTableHeader.tsx
