+++ packages/web/src/layouts/leaderboard-layout/components/leaderboard-table-row/LeaderboardTableRow.styles.ts
