+++ packages/web/src/layouts/leaderboard-layout/components/leaderboard-table-wrapper/LeaderboardTableWrapper.styles.tsx
