+++ packages/web/src/layouts/leaderboard-layout/components/leaderboard-table/mobile-leaderboard-table/MobileLeaderboardTable.tsx
