+++ packages/web/src/layouts/leaderboard-layout/components/leaderboard-table/tablet-leaderboard-table/TabletLeaderboardTable.tsx
