+++ packages/web/src/layouts/leaderboard-layout/components/learn-more/LearnMore.tsx
