+++ packages/web/src/layouts/leaderboard-layout/components/mobile-leaderboard-table-row/MobileLeaderboardTableRow.stories.tsx
