+++ packages/web/src/layouts/leaderboard-layout/components/next-update/NextUpdate.styles.ts
