+++ packages/web/src/layouts/leaderboard-layout/components/point-composition/PointComposition.tsx
