+++ packages/web/src/layouts/leaderboard-layout/containers/leaderboard-list-header-container/LeaderboardListHeaderContainer.styles.ts
