+++ packages/web/src/layouts/leaderboard-layout/containers/leaderboard-subheader-container/LeaderboardSubheaderContainer.tsx
@@ -1,4 +1,4 @@
-import LeaderboardSubheader from "@components/leaderboard/leaderboard-subheader/LeaderboardSubheader";
+import LeaderboardSubheader from "@layouts/leaderboard-layout/components/leaderboard-subheader/LeaderboardSubheader";
 import { useAddress } from "@hooks/address/use-address";
 
 export default function LeaderboardSubheaderContainer() {
