+++ packages/web/src/layouts/leaderboard-layout/containers/leaderboard-table-header-container/LeaderboardTableHeaderContainer.tsx
@@ -1,14 +1,15 @@
-import LeaderboardTableHeader from "@components/leaderboard/leaderboard-table-header/LeaderboardTableHeader";
 import { useWindowSize } from "@hooks/common/use-window-size";
 import {
   LEADERBOARD_TD_WIDTH,
   MOBILE_LEADERBOARD_TD_WIDTH,
   TABLET_LEADERBOARD_TD_WIDTH,
 } from "@constants/skeleton.constant";
+
+import LeaderboardTableHeader from "../../components/leaderboard-table-header/LeaderboardTableHeader";
 import {
   TABLE_HEAD,
   TABLE_HEAD_MOBILE,
-} from "@layouts/leaderboard-list-layout/LeaderboardListLayout";
+} from "../../leaderboard-list/LeaderboardList";
 
 export default function LeaderboardTableHeaderContainer() {
   const { isMobile, isTablet } = useWindowSize();
