+++ packages/web/src/layouts/leaderboard-layout/containers/leaderboard-table-skeleton-container/LeaderboardTableSkeletonContainer.tsx
@@ -1,11 +1,14 @@
 import TableSkeleton from "@components/common/table-skeleton/TableSkeleton";
-import LeaderboardTableWrapper from "@components/leaderboard/leaderboard-table-wrapper/LeaderboardTableWrapper";
 import {
   LEADER_INFO,
   MOBILE_LEADER_INFO,
   TABLET_LEADER_INFO,
 } from "@constants/skeleton.constant";
 import { useWindowSize } from "@hooks/common/use-window-size";
+
+import LeaderboardTableWrapper from "../../components/leaderboard-table-wrapper/LeaderboardTableWrapper";
+
+
 
 export default function LeaderboardTableSkeletonContainer() {
   const { isMobile, isTablet, isWeb } = useWindowSize();
