+++ packages/web/src/layouts/leaderboard-layout/leaderboard-list/LeaderboardList.styles.ts
